wrap json_object_unref with g_clear_pointer
There were errors in stderr about NULL being passed to
json_object_unref, so find all relevant places where this function is
called and use g_clear_pointer to ensure NULL is not passed to
json_object_unref.

There is one location in src/trg-client.c where some additional logic
is used after a NULL check on the pointer, so that is left the same.