add icons via gresouces.xml file
The gtk-connect, gtk-disconnect, and gtk-apply are all deprecated icons
in gtk3 and cannot be relied upon in future versions. However, they do
the best job for conveying the actual meaning of the actions they
represent in transmission-remote-gtk, so they should be vendored and
displayed as is. This is accomplished via a gresouces.xml file to
compile them to C code and bundle them at compile time.

Makefile and .gitignore changes are also included.