fix passing NULL to json_object_get_array_member
The trg_prefs_get_rss function was previously directly using
json_object_get_array_member and trying to return the array. The issue
is that the "rss" key didn't exist in the config, so it returned NULL.
Luckily, a function already exists to check this, and populate the array
if it doesn't exist: trg_prefs_get_array(). Since there is already a
null check on feeds, this works with minimal changes.