fix popup menus on wayland
gtk_menu_popup() is deprecated since it doesn't behave well on non-X11
displays. The replacement for this is gtk_menu_popup_at_pointer()
specifically when you want the menu to appear as a right-click menu.
This greatly simplifies the arguments. The NULL check is no longer
necessary because NULL is a valid argument for the event, it simply
needs to be cast back to a GdkEvent.