initialize cell height/width variables
Previously these values were undefined and there was a possibility of
reaching `h` without initializing it. Since this is undefined behavior,
set it to 0 as well as `w`. If `h` is reached while it is simply dropped
from the equation, most likely what was happened if it was uninitialized
as well