remove GtkStatusIcon code
GtkStatusIcon has been deprecated since 3.14, it was an implementation
of the legacy X11 systray functionality that is no longer relevant for
modern desktop environments. AppIndicators are the replacement on most
DEs, with GNOME being a notable exception. Though there is a gnome-shell
extension to add support.

Since transmission-remote-gtk already has AppIndicator support, only a
minor amount of refactoring was needed to drop support all together.
With libappindicator disabled, the menu options for the system tray will
no longer be visible in local preferences.