use GNotification instead of libnotify
This allows transmission-remote-gtk to drop a dependency and make it
easier to build. GNotification provides the same notification support
that libnotify provides, with the one exception being it cannot set an
explicit timeout. However, most desktop environments seem to ignore
this, and the timeout was hardcoded and not very user friendly to begin
with.