remove GtkAlignment
Deprecated in 3.14, replace with calling explicit xalign/yalign for
labels. This commit also removes hig_workarea_add_section_divider()
which is only used in the torrent props dialog. All this did was add a
tiny buffer between sections, unseen in any other properties dialog such
as local preferences. Since it serves very little purpose, is
arguably different than other windows, and uses an object that is
deprecaed to acheive the effect; the whole function was removed.