change GtkTable to GtkGrid
GtkTable was deprecated in 3.4, and GtkGrid is the replacement. This
required some refactoring due to the APIs not being 1:1. In particular,
setting GtkAttachOptions and x/ypadding in GtkGrid was not directly
possible. This was mostly worked around by clever tricks with the
underlying GtkLabel children in the grid itself. It mostly looks the
same, though the resizing certainly could be better.