cleanup pointer freeing in trg-client.c
This is an assorted bunch of fixes to use g_clear_pointer() whenever
reasonable in this file. This does not solve any bugs that I can see,
but previously this was a source of bugs causing GTK and JSON warnings,
so might as well do the whole file.

Additionally some whitespace was fixed.