unconditionally escape strings in add_links_to_text
In src/trg-general-panel.c comments are checked for links and then
marked up unconditonally. However, add_links_to_text() did not return a
markup-escaped string when a link was *not* found, and then passed the
raw string directly to gtk_label_set_markup(). This caused some comments
to generate warnings that the markup was invalid. Fix this by
unconditionally escaping comment strings.