Support torrent labels
This commit adds support for displaying labels from RPC as well as
adding them to a torrent via RPC.

Labels are checked via a regular expression to indicate if a label is
valid or not. If a label is not valid, you cannot add the torrent and
have to first fix the label string. An emptry string is considered valid
and will not add any labels.