Use clang-format for code style
This is highly opinionated, but it is a style that tries to reduce the
amount of churn in the code base. There are many places where whole
files need to be re-written because their styles are incredibly outside
of every other file.

The CI is updated with a format checker as well as an included
scripts/pre-commit git hook for local use.