Use libsoup3 over curl
libsoup is a gobject based API with async that integrates well into the
glib mainloop. The current curl threadpool implementation has a high
amount of latency and is somewhat complex, so this simplifies a lot of
code as well as adding a gobject based library for further extension.

The changes can be summarized as the following

1. remove the threadpool, and curl handling
2. Vastly simplify trg_request and make it private, as well as removing
   the ability to do anything but handle JSON RPC
3. Put a SoupSession into TrgClient and allow that to handle all
   connections to the daemon
4. Arrange a series of callbacks for async RPC as well as JSON loading
   of results (see comment in src/trg-client.c)
5. Set up new error handling for bad requests
6. Clean up the rest of the code base to work with the above changes.

Since this commit is very complex, even though it isn't 1:1 feature
complete with the old curl implementation, other things will be added as
separate commits for easier reviewing.

This commit additionally bumps glib to 2.70 for GUri usage

Co-Authored-By: Patrick Griffis <pgriffis@igalia.com>