Remove RSS support
RSS support is blocking a few simplifications for some other
improvements, and it additionally is very much untested and relies on
old, abandoned libraries for XML parsing.

Other technical factors influenced this decision (see github issue #191)
but the primary motivation is developer time and knowledge to maintain
this support. Frankly, there is none. Patches welcome.