Use dep.name() to find ayatana name
pkg-config variables are defined as `key=value`, and the `Key: Value`
versions are just metadata. Therefore most pkg-config files do not have
'Name' as a variable and `get_variable('Name')` will not work. However,
meson provides a way to find this with the dep.name() method, so use
that instead.