remove use of g_error() in torrent_add_dialog and base64encode
g_error is only supposed to be for unexpected errors, and therefore
should not be used for expected errors. In torrent-add-dialog, the
specific issue is that not having read permissions on a file is indeed
an *expected* error. In trg_base64encode, not being able to base64encode
a file for some reason is also expected. These errors should be passed
back to the user,

Some refactoring was needed to create a general error-dialog message.
There are also still unrelated g_error() usages that need to be cleaned
up.