reduce g_error() to g_warning() for trg_prefs_save()
In trg_prefs_constructor, g_get_user_config_dir() is used to create
priv->file. This value is always what XDG_CONFIG_HOME is, or at least
what the default is, e.g. ~/.config. There should be no circumstances,
outside of explicit user intervention, for this to not be user
writeable.

That being said, if a permissions issue existed, the code was set to
core dump via g_error(). That really shouldn't be the response to this
as there are cases where this could happen, though it is almost
certainly user error outside TRG's control. Therefore, bump this down to
a warning. If someone hits this, they'll be able to see what the warning
is and understand it.