Remove older-style GObject declarations
This is a massive commit that surely breaks *something* (but I don't
know what right now). Fundamentally all it does is switch to glib 2.70's
API and removes a ton of deprecated declarations and macros. It
accomplishes this by modifying how pretty much every single GObject
subclass used in this project is made. With this change comes a few
other things:

 1. #pragma once, if a header was changed
 2. Removing Privates in final types, just put the vars directly in the
    struct
 3. Since 2) was done, a lot of code was able to be removed that set up
    priv-> struct access.
 4. Renaming variables where appropriate due to 3) (via find/replace and
    sed)

Scattered throughout this is also a bunch of general cleanups I took the
liberty of doing when I saw them.