Fix windows CI's "allow dirty" fix with git config
The alternative fix would be opening a PR to meson project to do this change

```patch
-        ret = subprocess.call(['git', '-C', self.src_root, 'diff-index', '--quiet', 'HEAD'])
+        ret = subprocess.call(['git', '-C', self.src_root, 'diff-index', '--quiet', '--ignore-cr-at-eol', 'HEAD'])
```
in
https://github.com/mesonbuild/meson/blob/64535bbc821c0276ffe8c894502067a14229e10b/mesonbuild/mdist.py#L174-L180

but only in the cases where the host is windows and the git's `core.autocrlf` config is not set to `false`.

Which is more effort to me right now, than to patch this file. :3